+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: emaillet <emaillet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 14:52:01 by emaillet          #+#    #+#             */
-/*   Updated: 2024/10/15 19:46:41 by emaillet         ###   ########.fr       */
+/*   Updated: 2024/10/15 22:10:14 by emaillet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int					i;
+	size_t				i;
 	unsigned char		*local_dest;
 	unsigned const char	*local_temp;
